Move manifest-helpers to parent
